requirements.txt
flask
flask-mail
werkzeug
tinydb
python-dotenv
reportlab
qrcode[pil]
google-generativeai
pyopenssl
gspread==5.12.0
google-auth==2.25.2
authlib
PyPDF2
